moved get_source into journalist_app/utils.py
